+++ sdk/identity/identity/src/credentials/defaultAzureCredential.browser.ts
@@ -18,7 +18,6 @@ const logger = credentialLogger("DefaultAzureCredential");
  * Only available in NodeJS.
  */
 export class DefaultAzureCredential extends ChainedTokenCredential {
-  public static credentialStack: unknown[] = [];
   /**
    * Creates an instance of the DefaultAzureCredential class.
    *